# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ActressView:
    properties:
      aliases:
        type: string
      created_at:
        type: string
      dmm_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      thumb_url:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressTranslationView'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed:
    properties:
      error:
        example: Result not found in job
        type: string
      result_id:
        example: uuid-1
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeRequest:
    properties:
      result_ids:
        example:
        - uuid-1
        - uuid-2
        items:
          type: string
        type: array
    required:
    - result_ids
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeResponse:
    properties:
      excluded:
        items:
          type: string
        type: array
      failed:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed'
        type: array
      job:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse'
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult:
    properties:
      actress_sources:
        additionalProperties:
          type: string
        type: object
      candidates:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.ScrapeCandidate'
        type: array
      ended_at:
        type: string
      error:
        type: string
      field_sources:
        additionalProperties:
          type: string
        type: object
      file_path:
        type: string
      has_conflict:
        type: boolean
      is_multi_part:
        type: boolean
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
      movie_id:
        type: string
      part_number:
        type: integer
      part_suffix:
        type: string
      result_id:
        description: Stable UUID — survives movie_id changes
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus'
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse:
    properties:
      completed:
        type: integer
      completed_at:
        type: string
      destination:
        type: string
      excluded:
        additionalProperties:
          type: boolean
        type: object
      failed:
        type: integer
      files:
        items:
          type: string
        type: array
      id:
        type: string
      operation_count:
        type: integer
      operation_mode_override:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_operationmode.OperationMode'
      persist_error:
        type: string
      progress:
        type: number
      results:
        additionalProperties:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult'
        type: object
      reverted_count:
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus'
      total_files:
        type: integer
      update:
        type: boolean
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.DataSource:
    properties:
      confidence:
        description: Confidence score (0.0-1.0)
        example: 0.9
        type: number
      last_updated:
        description: ISO 8601 timestamp
        example: "2024-01-15T10:30:00Z"
        type: string
      source:
        description: '"scraper" or "nfo"'
        example: nfo
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.GenreView:
    properties:
      id:
        type: integer
      name:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreTranslationView'
        type: array
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.MergeStatistics:
    properties:
      conflicts_resolved:
        example: 5
        type: integer
      empty_fields:
        example: 2
        type: integer
      from_nfo:
        example: 3
        type: integer
      from_scraper:
        example: 10
        type: integer
      merged_arrays:
        example: 2
        type: integer
      total_fields:
        example: 15
        type: integer
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieResponse:
    properties:
      merge_stats:
        allOf:
        - $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MergeStatistics'
        description: Merge statistics when NFO merging occurred
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
      provenance:
        additionalProperties:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.DataSource'
        description: Field-level data source tracking
        type: object
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView:
    properties:
      actresses:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      director:
        type: string
      id:
        type: integer
      label:
        type: string
      language:
        type: string
      maker:
        type: string
      movie_id:
        type: string
      original_title:
        type: string
      series:
        type: string
      settings_hash:
        type: string
      source_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieView:
    properties:
      actresses:
        description: |-
          Relationships (contract DTOs — see collection_views.go; persistence
          models.* types never cross the API boundary)
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressView'
        type: array
      code:
        description: Canonical JAV code (e.g., "IPX-535"). Renamed from content_id.
        example: IPX-535
        type: string
      cover_url:
        type: string
      created_at:
        description: Audit timestamps
        type: string
      cropped_poster_url:
        type: string
      description:
        description: Metadata
        type: string
      director:
        type: string
      display_title:
        description: Titles
        example: IPX-535 Beautiful Woman
        type: string
      genres:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreView'
        type: array
      id:
        description: Secondary identifier — often equals Code but can differ.
        example: IPX-535
        type: string
      label:
        description: Sub-label
        type: string
      maker:
        description: Studio/maker
        type: string
      original_cover_url:
        type: string
      original_cropped_poster_url:
        type: string
      original_filename:
        type: string
      original_poster_url:
        type: string
      original_should_crop_poster:
        type: boolean
      original_title:
        example: 美人
        type: string
      poster_url:
        description: Poster / cover (flattened from PosterState — no custom marshaler
          needed)
        type: string
      rating_score:
        type: number
      rating_votes:
        type: integer
      rating_warning:
        type: string
      release_date:
        type: string
      release_year:
        type: integer
      runtime:
        description: in minutes
        type: integer
      screenshot_urls:
        items:
          type: string
        type: array
      series:
        description: Series name
        type: string
      should_crop_poster:
        type: boolean
      source_name:
        description: Source provenance
        type: string
      source_url:
        type: string
      title:
        example: Beautiful Woman
        type: string
      trailer_url:
        description: Media
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropRequest:
    properties:
      height:
        minimum: 1
        type: integer
      max_poster_height:
        description: |-
          MaxPosterHeight optional override for the max poster height (px). 0 = no cap.
          When omitted, the configured output.max_poster_height is used.
        minimum: 0
        type: integer
      width:
        minimum: 1
        type: integer
      x:
        minimum: 0
        type: integer
      "y":
        minimum: 0
        type: integer
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropResponse:
    properties:
      cropped_poster_url:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLResponse:
    properties:
      cropped_poster_url:
        type: string
      poster_url:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.UpdateMovieRequest:
    properties:
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
    required:
    - movie
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.Actress:
    properties:
      aliases:
        description: Pipe-separated
        type: string
      created_at:
        type: string
      dmm_id:
        description: Real DMM actress ID when available (unique only for values >
          0)
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      thumb_url:
        type: string
      translations:
        description: |-
          Translations carries optional translation data for API response enrichment.
          Not persisted in database — gorm:"-" prevents auto-migration/association.
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.ActressTranslation'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.ActressTranslation:
    properties:
      actress_id:
        type: integer
      created_at:
        type: string
      display_name:
        type: string
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      language:
        type: string
      last_name:
        type: string
      settings_hash:
        type: string
      source_name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.JobStatus:
    enum:
    - pending
    - running
    - completed
    - failed
    - cancelled
    - organized
    - reverted
    type: string
    x-enum-varnames:
    - JobStatusPending
    - JobStatusRunning
    - JobStatusCompleted
    - JobStatusFailed
    - JobStatusCancelled
    - JobStatusOrganized
    - JobStatusReverted
...
  github_com_javinizer_javinizer-go_internal_operationmode.OperationMode:
    enum:
    - organize
    - in-place
    - in-place-norenamefolder
    - metadata-artwork
    - preview
    type: string
    x-enum-varnames:
    - OperationModeOrganize
    - OperationModeInPlace
    - OperationModeInPlaceNoRenameFolder
    - OperationModeMetadataArtwork
    - OperationModePreview
...
  internal_api_actress.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  internal_api_actress.actressSyncCandidatesResponse:
    properties:
      actresses:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.Actress'
        type: array
      ids:
        items:
          type: integer
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8765
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/actresses/sync-candidates:
    get:
      description: Return IDs of actresses missing a DMM ID or profile thumbnail
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_actress.actressSyncCandidatesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_actress.ErrorResponse'
      summary: List actresses missing metadata
      tags:
      - actress
...
  /api/v1/batch/{id}/movies/batch-exclude:
    post:
      consumes:
      - application/json
      description: 'Exclude multiple movies from a batch job in a single request.
        Best-effort: excludes as many as possible and returns per-result failures.'
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result IDs to exclude
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Bulk exclude movies from batch organization
      tags:
      - web
...
  /api/v1/batch/{id}/results/{resultId}:
    patch:
      consumes:
      - application/json
      description: Update a movie's metadata within a batch job's results
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      - description: Updated movie data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.UpdateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Update movie in batch job
      tags:
      - web
  /api/v1/batch/{id}/results/{resultId}/exclude:
    post:
      description: Mark a movie in a batch job as excluded from file organization
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Exclude movie from batch organization
      tags:
      - web
...
  /api/v1/batch/{id}/results/{resultId}/poster-crop:
    post:
      consumes:
      - application/json
      description: Re-crop a temp poster for the review page using fixed-size crop
        coordinates
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      - description: Crop coordinates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Update manual poster crop in batch job
      tags:
      - web
  /api/v1/batch/{id}/results/{resultId}/poster-from-url:
    post:
      consumes:
      - application/json
      description: Download a poster image from a URL and set it as the movie's poster
        in the batch job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      - description: Poster URL
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Download poster from URL
      tags:
      - web
...
schemes:
- http
- https
swagger: "2.0"